Tests for BG run
These tests DO fail when used on the older bitmap code - the code that
generated the white lines.  It seems that the white lines were due to
incorrectly handing the first scanline "feature".

Seriously, whose idea was that?  Or was it a mistake, that had to be
hardened into the spec?