Move pressedKeys to RdesktopCanvas from Input
It's supposed to fix an alleged problem in some versions of Java where keys are not pressed, only released.  I'm not sure if this is still a problem, but the field is actually only used by RdesktopCanvas (with its KeyEvents).  For other library uses it doesn't make sense to have this field.